+++ CHANGELOG.md
@@ -4,6 +4,9 @@
 
 - Added support for publishing through the new [Central Portal](https://central.sonatype.com). To use
   this use the `CENTRAL_PORTAL` option when specifying the Sonatype host.
+- For Kotlin Multiplatform the main plugin will now automatically publish the
+  `release` variant if the project has an Android target.
+- Support specifying the Android variants to publish in `KotlinMultiplatform(...)`.
 - Updated minimum supported Gradle, Android Gradle Plugin and Kotlin versions.
 - Removed support for the deprecated Kotlin/JS plugin.
 - Removed the deprecated `closeAndReleaseRepository` task. Use `releaseRepository`, which